feat: add polling to the server
This commit adds polling to the server. Now, multipe clients can connect
at once and all will be served.
A problem, howerver, still exists. If a client begins to write something
but then does not complete it, the server will be stuck listening for
this client only, until it finishes what it was writing. This is not
very convenient.